feat: Adding Obstacle object to the Game , Solving bug
The game comes with 3 levels:
1- Easy: there is no obstacles.
2- Medium: the boarder of the window is an obstacle.
3- Hard: the boarder and '+' shape in the middle are obstacles.

if the snake touch the obstacle it will die.

The bug:
some times the food disapear from the window after the snake eat the food.

The reason:
the random enging place the food beyond the window.

The solution:
the random engine has start and end, the start is 0,
and the end is width of window. The end should be the width
of window - 1.